human resource template delete condition for added for used node in human resource

+++ module/GovtStakeholder/App/Http/Controllers/HumanResourceTemplateController.php
@@ -196,6 +196,13 @@ class HumanResourceTemplateController extends BaseController
     {
         $organizationUnitType = $humanResourceTemplate->organizationUnitType;
 
+        if ($humanResourceTemplate->humanResource->isNotEmpty()) {
+            return response()->json([
+                'message' => __('Non-deletable! Human resource using this node.'),
+                'alertType' => 'error',
+            ]);
+        }
+
         try {
             $humanResourceTemplate->delete();
         } catch (\Throwable $exception) {
